+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * [BUG] - [Researcher preview phone number/word missing](https://trello.com/c/Vt4JdtH5/293-bugs-researcher-preview)
 * [BUG] - [Live previews broken in every live environment](https://trello.com/c/BV2jgy9O/295-live-previews-broken-in-every-live-environment)
 * [BUG] - [Apply design rules featured in the original Barnardo's styleguide](https://trello.com/c/CtYloVeU/286-apply-design-rules-featured-in-the-original-barnardos-styleguide)
+* [FEATURE] - [Live preview: Recording](https://trello.com/c/YY9TDLiB/258-5-live-preview-recording)
 
 # 0.5.0 / 2017-11-16